Début back ArticleAVendre

+++ src/main/java/fr/eni/spring/Projet02/dal/CategorieDAOImpl.java
@@ -2,7 +2,6 @@ package fr.eni.spring.Projet02.dal;
 
 import fr.eni.spring.Projet02.bo.Categorie;
 import org.springframework.jdbc.core.BeanPropertyRowMapper;
-import org.springframework.jdbc.core.SqlInOutParameter;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 
@@ -14,7 +13,7 @@ public class CategorieDAOImpl implements CategorieDAO {
     private static final String FIND_ALL = "SELECT no_categorie,libelle FROM CATEGORIES";
     private static final String FIND_BY_ID = "SELECT no_categorie,libelle FROM CATEGORIES "+"WHERE no_categorie= :id";
     @Override
-    public Categorie read(int id) {
+    public Categorie read(long id) {
         MapSqlParameterSource parameterSource = new MapSqlParameterSource();
         parameterSource.addValue("id", id);
         return jdbcTemplate.queryForObject(FIND_BY_ID, parameterSource, new BeanPropertyRowMapper<>(Categorie.class));
